debugged and reorganized code
